refactor: [story-5] update tax
