+++ calTax/errorMessages.go
@@ -1,0 +1,7 @@
+package calTax
+
+var lowerTax = "Could you please check again, tax should getter than 0 or equal."
+var invalidProperty = "Allownance propperties does not valid"
+var badRequest = "Could you please check your request data agin"
+var faltalErrorPrepare = "can't prepare statment update"
+var faltalErrorExecute = "error execute update"
